Autogenerate and add pyi stub file for rhino3dm

+++ src/bindings/bnd_object.cpp
@@ -1426,7 +1426,7 @@ BND_CommonObject* BND_CommonObject::Decode(BND_DICT jsonObject)
   std::string decoded = base64_decode(buffer);
   int rhinoversion = IntFromDict(jsonObject,"archive3dm");
   int opennurbsversion = IntFromDict(jsonObject,"opennurbs");
-  int length = decoded.length();
+  int length = (int)decoded.length();
   const unsigned char* c = (const unsigned char*)&decoded.at(0);
   ON_Object* obj = ON_ReadBufferArchive(rhinoversion, opennurbsversion, length, c);
   return CreateWrapper(obj, nullptr);
